retrieval-doc/requirements.txt
fastapi
uvicorn
httpx
pdfplumber
pydantic
python-multipart
faiss-cpu
openai
tqdm